quadratic

+++ PrimeNumber/src/Elements/MyHelper.h
@@ -15,7 +15,7 @@ namespace Elements {
 		static void PrintMatrix(vector<vector<bool>> matrix);
 		static vector<vector<bool>> GetIdentityMatrix(unsigned long long n);
 		static void PowCExpD(mpz_t r, mpz_t c, mpz_t d);
-		static unsigned long GetUpperBoundOfPrimes(string n);
+		static long long GetUpperBoundOfPrimes(string n);
 	};
 }
